Create RUN_04-1.llsp
The alignment: the black and yellow lifter attachment is face forward and it is loaded with 3 energy units. Next, look at the lego piece at the back corner, if the holes are both in the middle of the 3rd and the 9th lines (count from the right). Lastly press it to the back of the field and lift to avoid swerving